pyproject.toml
[project]
name = "evidentai"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "cohere>=5.21.1",
    "datasets>=4.8.4",
    "langchain-text-splitters>=1.1.1",
    "openai>=2.30.0",
    "pymupdf>=1.27.2.2",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "qdrant-client>=1.17.1",
    "ragas>=0.4.3",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.56.0",
]
